Undo splitting DispatcherRegistry into an interface and an implementation base.
